+++ src/main/java/org/jamesframework/examples/clique/GreedyCliqueNeighbourhood2.java
@@ -20,11 +20,11 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Random;
 import java.util.Set;
-import java.util.concurrent.ThreadLocalRandom;
 import org.jamesframework.core.subset.SubsetSolution;
 import org.jamesframework.core.search.neigh.Move;
 import org.jamesframework.core.search.neigh.Neighbourhood;
 import org.jamesframework.core.subset.neigh.moves.AdditionMove;
+import org.jamesframework.core.util.Randomization;
 
 /**
  * Alternative implementation of greedy clique neighbourhood that uses the optimized clique solution type
@@ -46,7 +46,7 @@ public class GreedyCliqueNeighbourhood2 implements Neighbourhood<CliqueSolution>
         if(allMoves.isEmpty()){
             return null;
         } else {
-            Random rg = ThreadLocalRandom.current();
+            Random rg = Randomization.getRandom();
             return allMoves.get(rg.nextInt(allMoves.size()));
         }
     }